Renamed workspace tab back to general tab
