+++ src/main/java/org/jabref/gui/preferences/general/GeneralTab.fxml
@@ -15,8 +15,8 @@
 <?import org.jabref.gui.icon.JabRefIconView?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-         fx:controller="org.jabref.gui.preferences.workspace.WorkspaceTab">
-    <Label styleClass="titleHeader" text="%Workspace"/>
+         fx:controller="org.jabref.gui.preferences.general.GeneralTab">
+    <Label styleClass="titleHeader" text="%General"/>
     <GridPane hgap="10.0" vgap="4.0">
         <columnConstraints>
             <ColumnConstraints hgrow="SOMETIMES" percentWidth="30.0"/>
