+++ src/main/resources/l10n/JabRef_en.properties
@@ -958,8 +958,6 @@ Write\ metadata\ for\ all\ PDFs\ in\ current\ library?=Write metadata for all PD
 Writing\ metadata\ for\ selected\ entries...=Writing metadata for selected entries...
 Writing\ metadata...=Writing metadata...
 
-Workspace=Workspace
-
 Embed\ BibTeXEntry\ in\ PDF.=Embed BibTeXEntry in PDF.
 File\ '%0'\ is\ write\ protected.=File '%0' is write protected.
 Write\ BibTeXEntry\ as\ XMP\ metadata\ to\ PDF.=Write BibTeXEntry as XMP metadata to PDF.
